Update Trackball, cameratools, and camerautils to use dvec3/dmat4
- Trackball: All method signatures and internal computations use dvec3/dmat4
- cameratools.h: calculateFovBounds takes dvec3, perspectiveZoom uses dvec3/dmat3
- camerautils.h/cpp: setCameraView takes dmat3 view, internal fit functions use dvec3
- setCameraLookRanges uses dvec3 for camera property min/max values
- Fixed float/double conversions at boundaries (glm::distance -> static_cast)

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/2d668631-7819-4ed9-86d8-68721b62eaa8

+++ include/inviwo/core/algorithm/camerautils.h
@@ -91,8 +91,8 @@ IVW_CORE_API float fovxDegreesFrom(float fovyDegrees, float aspect);
  * @param updateLookRanges   the camera's look-to/look-from ranges are updated if Yes @see
  * setCameraLookRanges
  */
-IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBox, glm::mat3 view,
-                                float fitRatio = 1.05f,
+IVW_CORE_API void setCameraView(CameraProperty& cam, const glm::mat4& boundingBox,
+                                glm::dmat3 view, float fitRatio = 1.05f,
                                 UpdateNearFar updateNearFar = UpdateNearFar::No,
                                 UpdateLookRanges updateLookRanges = UpdateLookRanges::No,
                                 float maxZoomFactor = defaultZoomFactor,
